fix: add restore button to menu
When users create a new wallet and the click 'next' on the step where
they copy the seed phrase. If they close the extension before they enter
the password, they can not access the wallet.

By adding a `restore wallet` button to the menu, the user can easily
make the connection to restor the wallet with this button (granted the
seed phrase has been saved).