fix: cleanup keystores on popup disconnect
When a user enters the process to restore wallet from seed phrase,
the keystore is created even if they do not enter the password for the
account. The result was that the z3us menu could be filled up with
keystores that the user can not access.
This PR creates a clean up method for keystores that are not properly
setup, and removes them from the store, when the popup closes.